+++ GitSearch/GitSearch/Application/RepoInfo/RepoInfoCellView.swift
@@ -1,0 +1,20 @@
+//
+//  RepoInfoCellView.swift
+//  GitSearch
+//
+//  Created by Vldmr on 8/3/19.
+//  Copyright © 2019 mudritskiy. All rights reserved.
+//
+
+import UIKit
+
+class RepoInfoCellView: UITableViewCell {
+    
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
