+++ GitSearch/GitSearch/Application/SearchResult/SearchResultViewController.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-class HomeController: UICollectionViewController, UICollectionViewDelegateFlowLayout {
+class SearchResultViewController: UICollectionViewController, UICollectionViewDelegateFlowLayout {
     
     var gitRep: SearchInfo?
     var props = [String]()
@@ -17,7 +17,7 @@ class HomeController: UICollectionViewController, UICollectionViewDelegateFlowLa
         super.viewDidLoad()
         
         collectionView?.backgroundColor = UIColor.white
-        collectionView?.register(CellClass.self, forCellWithReuseIdentifier: "cellId")
+        collectionView?.register(SearchResultCellView.self, forCellWithReuseIdentifier: "cellId")
         
         let layout = UICollectionViewFlowLayout()
         layout.estimatedItemSize = CGSize(width: view.frame.size.width-32, height: 120)
@@ -40,7 +40,7 @@ class HomeController: UICollectionViewController, UICollectionViewDelegateFlowLa
     }
     
     override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "cellId", for: indexPath) as! CellClass
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "cellId", for: indexPath) as! SearchResultCellView
         cellProcessing(index: indexPath.row, cellInstance: cell)
         return cell
     }
@@ -89,7 +89,7 @@ class HomeController: UICollectionViewController, UICollectionViewDelegateFlowLa
         return result
     }
     
-    func cellProcessing(index: Int, cellInstance: CellClass? = nil) {
+    func cellProcessing(index: Int, cellInstance: SearchResultCellView? = nil) {
         guard let gitItems = self.gitRep?.items else { return }
         let item = strucToArray(resource: gitItems[index]) // why item's value are optional when type inside is non oprional String:String
         if let cell = cellInstance {
@@ -103,7 +103,7 @@ class HomeController: UICollectionViewController, UICollectionViewDelegateFlowLa
             """
         } else {
             // open cell
-            let newVC = CellDetailTableView()
+            let newVC = RepoInfoViewController()
             newVC.props = props
             props.forEach { prop in newVC.repInfo[prop] = item[prop] }
             navigationController?.pushViewController(newVC, animated: true)
@@ -111,75 +111,5 @@ class HomeController: UICollectionViewController, UICollectionViewDelegateFlowLa
     }
 }
 
-class CellClass: UICollectionViewCell, UICollectionViewDelegateFlowLayout {
-    
-    let title: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = UIFont.boldSystemFont(ofSize: 20)
-        label.textColor = UIColor.black
-        return label
-    }()
-    
-    let subTitle: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = UIFont.systemFont(ofSize: 11)
-        label.textColor = UIColor(red: 66/255, green: 85/255, blue: 99/255, alpha: 1.0)
-        label.numberOfLines = 0
-        return label
-    }()
-    
-    let subTitleBackground: UIView = {
-        let view = UIView()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        view.backgroundColor = UIColor(red: 232/255, green: 237/255, blue: 238/255, alpha: 1.0)
-        view.layer.cornerRadius = 5
-        return view
-    }()
-    
-    fileprivate func setupSubviews() {
-        contentView.addSubview(title)
-        contentView.addSubview(subTitleBackground)
-        contentView.addSubview(subTitle)
-        
-        let constraints = [
-            title.topAnchor.constraint(equalTo: contentView.topAnchor),
-            title.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-            title.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
-            //title.heightAnchor.constraint(equalToConstant: 16),
-            
-            subTitle.topAnchor.constraint(equalTo: title.bottomAnchor, constant: 16),
-            subTitle.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-            subTitle.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
-            
-            subTitleBackground.topAnchor.constraint(equalTo: subTitle.topAnchor, constant: -8),
-            subTitleBackground.leadingAnchor.constraint(equalTo: subTitle.leadingAnchor, constant: -8),
-            subTitleBackground.trailingAnchor.constraint(equalTo: subTitle.trailingAnchor, constant: 8),
-            subTitleBackground.bottomAnchor.constraint(equalTo: subTitle.bottomAnchor, constant: 8)
-        ]
-        NSLayoutConstraint.activate(constraints)
-        
-    }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-//        contentView.translatesAutoresizingMaskIntoConstraints = false
-        setupSubviews()
-    }
-    
-    override func preferredLayoutAttributesFitting(_ layoutAttributes: UICollectionViewLayoutAttributes) -> UICollectionViewLayoutAttributes {
-        setNeedsLayout()
-        layoutIfNeeded()
-        let size = contentView.systemLayoutSizeFitting(layoutAttributes.size)
-        var newFrame = layoutAttributes.frame
-        newFrame.size.height = ceil(size.height)
-        layoutAttributes.frame = newFrame
-        return layoutAttributes
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
+
