+++ GitSearch/GitSearch/Models/SearchItem.swift
@@ -1,18 +1,12 @@
 //
-//  CommonClasses.swift
+//  SearchItem.swift
 //  GitSearch
 //
-//  Created by Vldmr on 7/7/19.
+//  Created by Vldmr on 8/3/19.
 //  Copyright © 2019 mudritskiy. All rights reserved.
 //
 
 import Foundation
-
-class SearchInfo: Decodable {
-    let total_count: Int
-    let incomplete_results: Bool
-    let items: Array<SearchItem>?
-}
 
 class SearchItem: Decodable {
     let id: Int
