+++ GitSearch/GitSearch/Models/SpinnerViewController.swift
@@ -1,0 +1,27 @@
+//
+//  SpinnerViewController.swift
+//  GitSearch
+//
+//  Created by Vldmr on 8/3/19.
+//  Copyright © 2019 mudritskiy. All rights reserved.
+//
+
+import UIKit
+
+class SpinnerViewController: UIViewController {
+    var spinner = UIActivityIndicatorView(style: UIActivityIndicatorView.Style.gray)
+    
+    override func loadView() {
+        view = UIView()
+        //        view.backgroundColor = UIColor(white: 0, alpha: 0.3)
+        
+        spinner.translatesAutoresizingMaskIntoConstraints = false
+        spinner.startAnimating()
+        view.addSubview(spinner)
+        
+        NSLayoutConstraint.activate([
+            spinner.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            spinner.centerYAnchor.constraint(equalTo: view.centerYAnchor)
+            ])
+    }
+}
